+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 # Unreleased: pdoc next
 
+ - 
+
+# 2021-03-10: pdoc 6.4.0
+
  - Functions in the current scope can now be referenced without specifying
    the full qualified name. For example, one can use `bar()` instead of 
    `Foo.bar()` in the docstring of `Foo`.
- - Improve reStructuredText rendering: Add support for footnotes.
- - Improve reStructuredText rendering: Minor bugfixes.
+ - Numpydoc: *See Also* sections are now parsed properly.
+ - reStructuredText: Add support for footnotes and fix minor bugs.
 
 # 2021-02-24: pdoc 6.3.2